抽离登陆请求
