+++ system-service/src/main/java/com/ifox/platform/adminuser/rest/AdminUserController.java
@@ -1,20 +1,14 @@
 package com.ifox.platform.adminuser.rest;
 
-import com.ifox.platform.adminuser.exception.NotFoundAdminUserException;
-import com.ifox.platform.adminuser.exception.RepeatedAdminUserException;
-import com.ifox.platform.adminuser.request.LoginRequest;
 import com.ifox.platform.adminuser.request.SaveRequest;
 import com.ifox.platform.adminuser.response.AdminUserVO;
 import com.ifox.platform.adminuser.service.AdminUserService;
 import com.ifox.platform.common.rest.BaseResponse;
 import com.ifox.platform.common.rest.MultiResponse;
-import com.ifox.platform.common.rest.TokenResponse;
 import com.ifox.platform.entity.adminuser.AdminUserEO;
 import com.ifox.platform.utility.common.DigestUtil;
 import com.ifox.platform.utility.common.EncodeUtil;
-import com.ifox.platform.utility.common.ExceptionUtil;
 import com.ifox.platform.utility.common.PasswordUtil;
-import com.ifox.platform.utility.jwt.JWTHeader;
 import com.ifox.platform.utility.jwt.JWTUtil;
 import com.jsoniter.JsonIterator;
 import io.swagger.annotations.Api;
@@ -28,18 +22,14 @@ import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.ifox.platform.common.constant.RestStatusConstant.*;
+import static com.ifox.platform.common.constant.RestStatusConstant.SUCCESS;
 
-/**
- * 后台用户管理接口
- */
-@Api("后台用户管理接口")
+@Api(description = "后台用户管理", basePath = "/")
 @Controller
-@RequestMapping(value = "/adminUser", headers = "api-version=1.0")
+@RequestMapping(value = "/adminUser", headers = {"api-version=1.0", "Authorization"})
 public class AdminUserController {
 
     private Logger logger = LoggerFactory.getLogger(getClass());
@@ -50,45 +40,6 @@ public class AdminUserController {
     @Autowired
     private AdminUserService adminUserService;
 
-    @ApiOperation(value = "后台用户登录", notes = "后台用户登录接口")
-    @RequestMapping(value = "/login", method = RequestMethod.POST)
-    @ResponseBody
-    TokenResponse login(@ApiParam @RequestBody LoginRequest loginRequest){
-        logger.info("用户登陆:{}", loginRequest);
-        Boolean validAdminUser = false;
-        TokenResponse tokenResponse = new TokenResponse();
-        try {
-            validAdminUser = adminUserService.validLoginNameAndPassword(loginRequest.getLoginName(), loginRequest.getPassword());
-        } catch (NotFoundAdminUserException | RepeatedAdminUserException e) {
-            logger.error(ExceptionUtil.getStackTraceAsString(e));
-            tokenResponse.setStatus(NOT_FOUND);
-            tokenResponse.setDesc("用户不存在");
-            logger.info("登陆异常 loginName:{}", loginRequest.getLoginName());
-            return tokenResponse;
-        }
-
-        if (!validAdminUser) {
-            tokenResponse.setStatus(USER_NAME_OR_PASSWORD_ERROR);
-            tokenResponse.setDesc("用户名或者密码错误");
-            logger.info("用户名或者密码错误 loginName:{}", loginRequest.getLoginName());
-            return tokenResponse;
-        }
-
-        String secret = env.getProperty("jwt.secret");
-        try {
-            tokenResponse.setStatus(SUCCESS);
-            tokenResponse.setDesc("登陆成功");
-            String token = JWTUtil.generateJWT(new JWTHeader(), adminUserService.generatePayload(loginRequest.getLoginName()), secret);
-            tokenResponse.setToken(token);
-            logger.info("登陆成功 loginName:{}, token:{}", loginRequest.getLoginName(), token);
-        } catch (UnsupportedEncodingException e) {
-            tokenResponse.setStatus(SERVER_EXCEPTION);
-            tokenResponse.setDesc("服务器异常");
-            logger.error(ExceptionUtil.getStackTraceAsString(e));
-            logger.info("登陆异常 loginName:{}", loginRequest.getLoginName());
-        }
-        return tokenResponse;
-    }
 
     @ApiOperation(value = "保存用户信息")
     @RequestMapping(value = "/save", method = RequestMethod.POST)
@@ -116,6 +67,20 @@ public class AdminUserController {
         baseResponse.setStatus(SUCCESS);
         baseResponse.setDesc("保存成功");
         logger.info("保存成功:{}", saveRequest.getLoginName());
+        return baseResponse;
+    }
+
+    @ApiOperation(value = "删除用户")
+    @RequestMapping(value = "/delete", method = RequestMethod.DELETE)
+    @ResponseBody
+    BaseResponse delete(@ApiParam @RequestBody String[] userId){
+        logger.info("删除用户:{}", userId);
+        BaseResponse baseResponse = new BaseResponse();
+
+
+        baseResponse.setStatus(SUCCESS);
+        baseResponse.setDesc("删除成功");
+        logger.info("删除成功");
         return baseResponse;
     }
